final git

+++ src/test/testlog/TestFS.java
@@ -71,6 +71,9 @@ class TestFS {      //Tarei Marcello
     public boolean messageDaoFsDelete(){
         DAOMessageImplFileSystem msgDao = new DAOMessageImplFileSystem();
         Message msg = new Message(29, 0, 0, "", new Timestamp(System.currentTimeMillis()), 0);
+        if(msgDao.get(29)==null){
+            return false;
+        }
         msgDao.delete(msg);
         return true;
     }
